support more click objects
