+++ src/rich_click/cli.py
@@ -19,7 +19,7 @@ from rich.text import Text
 
 from rich_click import command as rich_command
 from rich_click import group as rich_group
-from rich_click import RichCommand, RichGroup
+from rich_click import RichBaseCommand, RichCommand, RichGroup, RichMultiCommand
 from rich_click.rich_click import (
     ALIGN_ERRORS_PANEL,
     ERRORS_PANEL_TITLE,
@@ -68,6 +68,8 @@ def patch() -> None:
     click.command = rich_command
     click.Group = RichGroup
     click.Command = RichCommand
+    click.BaseCommand = RichBaseCommand
+    click.RichMultiCommand = RichMultiCommand
 
 
 def main(args: Optional[List[str]] = None) -> Any:
